+++ rvp/builtin-rvp-smalxds.c
@@ -7,7 +7,7 @@
 
 #if defined(__riscv_zpsf) && !defined(__riscv_32e)
 static __attribute__ ((noinline))
-int64_t smalxds (int64_t t, unsigned long a, unsigned long b)
+int64_t smalxds (int64_t t, uint32_t a, uint32_t b)
 {
   return __rv__smalxds (t, a, b);
 }
