+++ rvp32_scan/builtin-rvp-smalda.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-int64_t adlams (int64_t t, unsigned long a, unsigned long b)
+int64_t adlams (int64_t t, uint32_t a, uint32_t b)
 {
   return __rv__smalda (t, a, b);
 }
